# docs/swagger.yaml
basePath: /v1
definitions:
...
  pkgs.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  user.AssignRolesToUserReq:
    properties:
      role_ids:
        description: 角色ID列表
        items:
          type: string
        minItems: 1
        type: array
    required:
    - role_ids
    type: object
  user.CreateUserReq:
    properties:
      password:
        type: string
      phone:
        maxLength: 11
        minLength: 11
        type: string
      profile:
        $ref: '#/definitions/user.Profile'
      username:
        type: string
    required:
    - password
    - phone
    - username
    type: object
  user.Profile:
    additionalProperties: true
    type: object
  user.UpdateUserReq:
    properties:
      password:
        description: 密码，前端进行MD5加密
        type: string
      profile:
        allOf:
        - $ref: '#/definitions/user.Profile'
        description: 个人信息
    type: object
  user.UserListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/user.UserRes'
        type: array
      total:
        type: integer
    type: object
  user.UserRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      profile:
        $ref: '#/definitions/user.Profile'
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support
  description: This is a sample server for go-pg-demo server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Go-PG Demo API
  version: "1.0"
paths:
...
  /user:
    post:
      consumes:
      - application/json
      description: 通过提供用户名、手机号、密码等信息创建一个新的用户账户。成功后返回新创建用户的唯一标识符(UUID)。
      parameters:
      - description: 创建用户所需的请求体参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建用户，返回用户ID
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数验证失败或格式不正确
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法创建用户
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 创建一个新的用户账户
      tags:
      - 用户管理
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: 通过指定的用户唯一标识符(UUID)来删除特定用户。这是一个永久性操作，请谨慎使用。
      parameters:
      - description: 用户唯一标识符(UUID)
        format: UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除用户，返回受影响的行数
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
        "400":
          description: 提供的用户ID格式无效
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法删除用户
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 根据用户ID删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 通过指定的用户唯一标识符(UUID)来检索特定用户的详细信息。
      parameters:
      - description: 用户唯一标识符(UUID)
        format: UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户信息
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserRes'
              type: object
        "400":
          description: 提供的用户ID格式无效
          schema:
            $ref: '#/definitions/pkgs.Response'
        "404":
          description: 未找到指定ID的用户
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法获取用户信息
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 根据用户ID获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 通过指定的用户唯一标识符(UUID)来更新特定用户的密码和个人信息。只会更新请求中包含的字段。
      parameters:
      - description: 用户唯一标识符(UUID)
        format: UUID
        in: path
        name: id
        required: true
        type: string
      - description: 需要更新的用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新用户信息
          schema:
            $ref: '#/definitions/pkgs.Response'
        "400":
          description: 请求参数验证失败或格式不正确
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法更新用户信息
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 根据用户ID更新用户信息
      tags:
      - 用户管理
  /user/{id}/role:
    post:
      consumes:
      - application/json
      description: 为指定用户分配一个或多个角色。该操作会完全替换用户当前的所有角色关系。
      parameters:
      - description: 用户唯一标识符(UUID)
        format: UUID
        in: path
        name: id
        required: true
        type: string
      - description: 要分配给用户的角色ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.AssignRolesToUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功为用户分配角色
          schema:
            $ref: '#/definitions/pkgs.Response'
        "400":
          description: 请求参数验证失败或格式不正确
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法为用户分配角色
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 为用户分配角色
      tags:
      - 用户管理
  /user/list:
    get:
      consumes:
      - application/json
      description: 获取系统中的用户列表，支持按手机号模糊搜索，并提供分页功能。
      parameters:
      - default: 1
        description: 页码，从1开始计算
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 每页条目数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: 手机号模糊搜索关键字
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户列表
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.UserListRes'
              type: object
        "400":
          description: 请求参数验证失败或格式不正确
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误，无法获取用户列表
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 获取用户列表（支持分页和筛选）
      tags:
      - 用户管理
securityDefinitions:
  JWT:
    description: JWT token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"